// src/contracts/generated/csi07/StrongActionDelegationRequestV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserAccountId } from "./BrowserAccountId";
import type { BrowserHeadReferenceV1 } from "./BrowserHeadReferenceV1";
import type { CatalogId } from "./CatalogId";
import type { Digest32 } from "./Digest32";

export type StrongActionDelegationRequestV1 = { schemaVersion: number, head: BrowserHeadReferenceV1, browserAccountId: BrowserAccountId, attemptId: string, operationKey: CatalogId, targetDigest: Digest32, };