// src/contracts/generated/csi07/StrongActionDelegationResultV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Origin } from "./Origin";
import type { Secret32 } from "./Secret32";

export type StrongActionDelegationResultV1 = { schemaVersion: number, identityApiOrigin: Origin, capability: string, csrf: Secret32, expiresAt: string, };